requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.7.4
numpy==1.24.4
pandas==2.0.3
Pillow==10.1.0
Pillow==10.2.0
plotly==5.18.0
PyAudio==0.2.11
pydub==0.25.1
Requests==2.31.0
streamlit==1.29.0
websockets==12.0
